Update TwoFactorAuth library
This includes PHP 8.5 deprecation fixes as well as a major update to v3 of the library.

You might want to also consider adding: https://github.com/endroid/qr-code based on: https://github.com/RobThree/TwoFactorAuth/issues/104

I haven't throughly tested this, but my changes to this module's files allow for setting up a new TOTOP and logging in  with it.